requirements.txt
astroid==2.3.2
docutils==0.15.2
isort==4.3.21
kafka-python==1.3.4
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
psutil==5.6.4
pylint==2.4.3
python-daemon==2.2.4
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
